requirements.txt
face_recognition==1.3.0
opencv-python==4.8.1
Pillow==10.1.0